m-fleet-ledger project:     created User Controller and service

+++ src/main/java/com/fleet/ledger/app/repository/UserAddressRepository.java
@@ -3,5 +3,6 @@ package com.fleet.ledger.app.repository;
 import com.fleet.ledger.app.entity.UserAddress;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserAddressRepository extends JpaRepository<UserAddress, Long> {
+public interface UserAddressRepository extends JpaRepository<UserAddress, Integer> {
+
 }
